Important Push incoming

+++ core/src/com/project_1_2/group_16/bot/ai/Neighbour.java
@@ -24,7 +24,7 @@ public class Neighbour extends Agent {
     }
 
 
-    public double getFitness(){
+    public float getFitness(){
         return this.fitness;
     }
